requirements.txt
selenium==4.16.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
mysqlclient==2.2.1